Edited signature for kstrlen

+++ kernel/libk/include/kstring.h
@@ -5,6 +5,6 @@
 #ifndef __K_STR_H__
 #define __K_STR_H__
 
-int kstrlen(const char *__str);
+size_t kstrlen(const char *__str);
 
 #endif // __K_STR_H__
